refactor: address code review findings on declaration PDF
- Fix barrel import violations: route and PDF document now import from
  module barrels instead of internal sub-paths
- Split DeclarationPdfDocument.tsx (526 lines) into separate section
  components under sections/
- Extract shared styles to pdfStyles.ts with named style entries
  replacing inline style objects
- Unify duplicate INFO_BG/TABLE_HEADER_BG into BLUE_FRANCE_LIGHT
- Deduplicate step4/step5 category mapping with mapStepCategories helper
- Remove unnecessary "use client" from DownloadDeclarationPdfButton
- Extract renderQuartileTable closure into QuartileTable component
