+++ packages/app/src/modules/declaration-remuneration/index.ts
@@ -2,13 +2,20 @@ export { DeclarationIntro } from "./DeclarationIntro";
 export { DeclarationLayout } from "./DeclarationLayout";
 export { MissingSiret } from "./MissingSiret";
 export { StepPageClient } from "./StepPageClient";
+export {
+	computeGap,
+	computePercentage,
+	formatCurrency,
+	formatGap,
+	gapLevel,
+} from "./shared/gapUtils";
 export { Step1Workforce } from "./steps/Step1Workforce";
 export { Step2PayGap } from "./steps/Step2PayGap";
 export { Step3VariablePay } from "./steps/Step3VariablePay";
 export { Step4QuartileDistribution } from "./steps/Step4QuartileDistribution";
 export { Step5EmployeeCategories } from "./steps/Step5EmployeeCategories";
 export { Step6Review } from "./steps/Step6Review";
-
+export { parseStep5Categories } from "./steps/step6/parseStep5Categories";
 export type {
 	CategoryData,
 	PayGapRow,
