+++ packages/app/src/modules/declarationPdf/index.ts
@@ -1,1 +1,3 @@
+export { buildPdfData } from "./buildPdfData";
+export { DeclarationPdfDocument } from "./DeclarationPdfDocument";
 export { DownloadDeclarationPdfButton } from "./DownloadDeclarationPdfButton";
